requirements-dev.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
pytest
pytest-cov
pytest-django
model-bakery
pylint
pylint-django
isort
black
